add: digest auth
