+++ app.py
@@ -1,9 +1,10 @@
 from flask import Flask, render_template, jsonify, request
-import pymysql
 import configparser
 import werkzeug
 from datetime import datetime
 import os
+from flask_httpauth import HTTPDigestAuth
+from models import fetchUsers, fetchAllFileData, saveUploadedFile
 
 
 # create custom flask instance
@@ -33,47 +34,33 @@ config.read('config.cfg')
 app.config['MAX_CONTENT_LENGTH'] = 1 * 1024 * 1024
 UPLOAD_DIR = './static/files'
 
+# auth settings
+auth = HTTPDigestAuth()
+users = fetchUsers()
+app.config['SECRET_KEY'] = 'test'
 
-def connectDb():
-    # 設定ファイルからDB接続情報を読み込み
-    host = config.get('development', 'MYSQL_DATABASE_HOST')
-    user = config.get('development', 'MYSQL_DATABASE_USER')
-    password = config.get('development', 'MYSQL_DATABASE_PASSWORD')
-    db = config.get('development', 'MYSQL_DATABASE_DB')
 
-    return pymysql.connect(
-        host=host,
-        user=user,
-        password=password,
-        db=db,
-        charset='utf8mb4',
-        cursorclass=pymysql.cursors.DictCursor
-    )
+@auth.get_password
+def get_pw(username):
+    print(users)
+    if username in users:
+        return users.get(username)
+    return None
 
 
 # route settings
 # ルーティング処理はVueRouterで制御するのでflask側では全てのリクエストにて対しindex.htmlを返す
 @app.route('/', defaults={'path': ''})
 @app.route('/<path:path>')
+@auth.login_required
 def index(path):
     return render_template('index.html')
 
 
 # api settings
 @app.route('/api/files', methods=['GET'])
-def fetchAllFileData():
-    dbh = connectDb()
-    cursor = dbh.cursor()
-    sql = 'select f.id as id, data, '\
-        'title, u.name as author, '\
-        'f.created_at as created_at '\
-        'from files as f join users as u '\
-        'on f.uploader_id = u.id'
-    cursor.execute(sql)
-    files = cursor.fetchall()
-    cursor.close()
-    dbh.close()
-
+def getAllFileData():
+    files = fetchAllFileData()
     return jsonify(files)
 
 
@@ -93,19 +80,6 @@ def uploadFile():
     file.save(saveFilePath)
     saveUploadedFile(saveFilePath)
     return 'result: upload is succeeded.'
-
-
-def saveUploadedFile(saveFilePath):
-    connection = connectDb()
-    try:
-        with connection.cursor() as cursor:
-            sql = 'insert into files(data, title, uploader_id)' \
-                ' values(%s, %s, %s)'
-            cursor.execute(sql, (saveFilePath, 'test', 1))
-            connection.commit()
-            cursor.close()
-    finally:
-        connection.close()
 
 
 @app.errorhandler(werkzeug.exceptions.RequestEntityTooLarge)
